requirements.txt
paho-mqtt>=1.2
humanfriendly>=2.1
requests>=2.23.0
